+++ appcues/src/main/java/com/appcues/data/mapper/step/primitives/ButtonPrimitiveMapper.kt
@@ -1,0 +1,12 @@
+package com.appcues.data.mapper.step.primitives
+
+import com.appcues.data.mapper.step.mapPrimitive
+import com.appcues.data.mapper.styling.mapComponentStyle
+import com.appcues.data.model.ExperiencePrimitive.ButtonPrimitive
+import com.appcues.data.remote.response.step.primitive.PrimitiveResponse.ButtonPrimitiveResponse
+
+internal fun ButtonPrimitiveResponse.mapButtonPrimitive() = ButtonPrimitive(
+    id = id,
+    content = content.mapPrimitive(),
+    style = style.mapComponentStyle(),
+)
