+++ appcues/src/main/java/com/appcues/data/mapper/step/primitives/ImagePrimitiveMapper.kt
@@ -1,0 +1,17 @@
+package com.appcues.data.mapper.step.primitives
+
+import com.appcues.data.mapper.styling.mapComponentContentMode
+import com.appcues.data.mapper.styling.mapComponentSize
+import com.appcues.data.mapper.styling.mapComponentStyle
+import com.appcues.data.model.ExperiencePrimitive.ImagePrimitive
+import com.appcues.data.remote.response.step.primitive.PrimitiveResponse.ImagePrimitiveResponse
+
+internal fun ImagePrimitiveResponse.mapImagePrimitive() = ImagePrimitive(
+    id = id,
+    url = imageUrl,
+    accessibilityLabel = accessibilityLabel,
+    style = style.mapComponentStyle(),
+    intrinsicSize = intrinsicSize.mapComponentSize(),
+    contentMode = mapComponentContentMode(contentMode),
+    blurHash = blurHash,
+)
