+++ appcues/src/main/java/com/appcues/data/remote/adapters/StepContainerAdapter.kt
@@ -1,9 +1,9 @@
 package com.appcues.data.remote.adapters
 
 import com.appcues.data.remote.response.action.ActionResponse
-import com.appcues.data.remote.response.step.primitive.PrimitiveResponse
 import com.appcues.data.remote.response.step.StepContainerResponse
 import com.appcues.data.remote.response.step.StepResponse
+import com.appcues.data.remote.response.step.primitive.PrimitiveResponse
 import com.appcues.data.remote.response.trait.TraitResponse
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.JsonClass
